First commit of MutationSpiker and ReplaceReads.
MutationSpiker is used to spike in mutation, simple deletions or simple insertions in reads and write the mutated reads to a BAM file. ReplaceReads is used to consume this bam file and the original bam file, to replace original reads with the mutated alternatives. **Note** CIGAR strings of reads that has had insertions or deletions spiked into them will remain being the CIGAR from the original BAM, and therefore incorrect. Therefore, the final BAM needs to be converted to fastq using for example picard SamToFastq.